% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptax-package.R
\docType{package}
\name{ptax-package}
\alias{ptax}
\alias{ptax-package}
\title{ptax: Disponibiliza a cotacao diaria da PTAX de 2010 a 2018}
\description{
Esse pacote prove a cotacao diaria da PTAX do Banco Central do Brasil de 2010 a 2018. Consiste na media das cotacoes de compra e venda de dolar informadas pelos operadores de cambio ao Banco Central do Brasil.
}
\keyword{internal}
