% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptax_diaria.R
\name{ptax_diaria}
\alias{ptax_diaria}
\title{PTAX Diaria}
\usage{
ptax_diaria(ano = NULL, mes = NULL, dia = NULL)
}
\arguments{
\item{ano}{YYYY (quatro digitos)}

\item{mes}{M ou MM (um ou dois digitos)}

\item{dia}{D ou DD (um ou dois digitos)}
}
\value{
Vetor com a data, o valor de compra e venda da ptax no ultimo dia util do mes
}
\description{
Retorna um vetor com os valores de compra e venda da PTAX na data solicitada.
Se for um dia nao util, nao existe cotacao PTAX, portanto retorna NULL
}
\examples{

ptax_diaria(2010,1,5)

ptax_diaria(2010,1,1)

}
