% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptax_mensal.R
\name{ptax_mensal}
\alias{ptax_mensal}
\title{PTAX Mensal}
\usage{
ptax_mensal(ano, mes)
}
\arguments{
\item{ano}{YYYY (quatro digitos)}

\item{mes}{M ou MM (um ou dois digitos)}
}
\value{
Vetor com a data, o valor de compra e venda da ptax no ultimo dia util do mes
}
\description{
Retorna um vetor com os valores de compra e venda da PTAX no ultimo dia util do mes,
definida como a ptax_mensal
}
\examples{

ptax_mensal(2010,1)

ptax_mensal(2010,13)

}
